refactor: post.ts to postDetail.ts
